Clean up B fields
There was some messiness between the B field transport options,
which has been addressed a bit: now there are basically just
some "do we have a field at all" flags in GRMHD package, and
a dummy B field allocated automatically rather than a "none"
option.

B_CD is still not a great option for spherical KS sims, but
this appears to be a consequence of the inner boundary
condition mostly.  If that can be addressed, KHARMA is
semi-usable for AMR simulations without further changes.

Note the Parthenon update breaks restarts and possibly
other things, this is not overall a production state of
the code.