Put variables on equal footing
One of the core parts of working with Parthenon is taking
the various named variables which may be defined by different
packages, and bundling them together to be accessed like a single
array. This can even be done over mesh blocks, which avoids kernel
launch overhead and is key to better strong scaling behavior in
KHARMA.

This commit standardizes a struct, VarMap in types.h, for recording
which variable is at which location in the pack, which can be used
on the device to access variables in much the same way as iharm3d's
macros, or KHARMA's enumerated value (the "prims::rho" stuff).

The difference is that it can be constructed independently for each
pack, allowing additional variables to be specified in arbitrary
combinations at run time.

With this struct, it uses packs in several more locations, including
a pack over meshblocks when calculating divB.

Also:
* Get rid of some hacky file-scope variables
* Standardize problem initialization routines
* Move wind term and current (jcon) caclculations to packages
* Allow combining (good for strong scaling & some GPUs) or
  splitting (good for vectorization) several more loops,
  via compile-time options.
* Script to update restart files from old versions of KHARMA,
  and fix Parthenon's mistaken MeshBlockSize accounting
* Option to add a "hot spot" of extra density when restarting.

Fix:
* Fix a recently resurfaced bug in boundary conditions when using
  Parthenon's special boundary for loops, by reverting to old loop
  bounds.