Resize meshes on restart
Updates the IharmRestart initialization so it can linearly interpolate
primitives onto a mesh given in its initialization file.  Since this
is now the primary reason to use it, the problem parameters have been
renamed to resize_restart for filename clarity.

Linear interpolation at domain boundaries could be better, and
initialization of ghost zones hasn't been well tested.  That said,
there's no issue resizing & restarting various torii.

Resizing a magnetic field requires revising B so that divB~=0. This is
done with a new package, "B_Cleanup", which implements a damped Jacobi
iterative solver for a scalar potential p, the divergence of which can
be subtracted from B to eliminate divB. This solver is implemented such
that it could be used every so often during a normal run, to cut down
on B field divergence that may have built up over ~millions of steps.

This solver is quite slow, costing the same as ~100M of evolution at
the new resolution.  I don't know if this is inevitable from problem
size, a consequence of the higher-stencil Laplacian we use for corner-
centered divB, or something about my implementation.  If the MPI syncs
could be restricted to just P, that might speed things up considerably.

This is on 'kharmaim' branch as it might have and might in near future
use the Kokkos kernels.