Faster resizing with BiCGStab
This is the payoff for the big commit of updates, which rebased to a
new branch of Parthenon.
This commit uses the BiCGStab solver introduced in that Parthenon branch
to replace the old damped Jacobi iteration when eliminating the magnetic
field divergence, e.g. after resizing.  It is *much* faster, going from
>48 hours for large resizes to about 5 minutes.
It also fixes one issue where resizing code would use the "startx" value
from the *new* grid when interpolating the *old* one, leading to an
outward shift for all data.