Prolongation operator.
Arrange flags to split physics conserved vs. code conserved vars.
(Better flag organization forthcoming which will hide some
silliness here.)
In particular, we want the code to conserve the face B,
but the cell-centered version is still "conserved" in the
physical sense.

Add the prolongation operator from Olivares et al. 2019 in
generality, but currently hardcoded to alpha values matching
the operator in Toth & Roe.

Boundaries between refinement levels are still a problem,
but overall this is pretty close for a day of fiddling.