First-Order Flux Corrections
This adds basic & perhaps quite slow FOFC to KHARMA.

The idea of FOFC is to, instead of directly adding material when
zones reach unphysical states, instead be proactive by reducing
the order of incoming fluxes, in the hope of a more accurate
and stable result than direct manipulation.

When a zone would violate any of the normal floors (though the
criterion could be more flexible), any bordering fluxes are
replaced with donor-cell, LLF values.  This is done locally
and before any flux adjustments, so it does not require
complex synchronization.

In basic testing the results are remarkable. While FOFC
(at least as implemented here) has a tendency to
"hallucinate" a tiny bit of structure at low density,
I can retain the relevant pieces of the Komissarov '99
explosion solution in some detail, using:
1. "linearized" weno5 reconstruction, HLLE fluxes
2. G&S '05 full upwinded EMF values
3. bsq_over_rho_max = 1e4, no gamma ceiling
4. cfl = 0.9
5. gamma = 5/3

It's pretty reasonable to expect this to allow eliminating the gamma
ceiling in tori, in favor of using b^2/rho to limit velocity, and
looking like there's very little stability penalty for enabling all
the new toys I've been adding...