Kastaun in GR
This migrates the Kastaun implementation to Phoebus's
general-coordinate variable prep (AthenaK's is specific
to gdet=1 and generalizing it was proving buggy).
It now passes at least a Bondi problem in 2D in MKS.
Using Kastaun will allow algorithm simplifications in a
bunch of places, many now done or in TODOs.

This also bundles a bunch of previous stuff hard to disentangle:
* Floor changes as needed since new inversion applies floors
* FOFC fixes, option to use GLF fluxes
* Fluxes: not as many silent sound speed limits & fixes
* Fallback reconstruction now PPM
* Flag printing fix