B Cleanup for Faces
Mainly this PR adds a laplacian for face-centered B to `CleanupDivergence`,
which is automatically used if the B_CT package is loaded.  It doesn't actually
seem faster than the cell-centered version despite having a smaller stencil.
It also adds a "PtoU" for face-centered fields, which takes a cell-centered B
and interpolates it to faces.  Yes, this means interpolating interpolated values,
which is not ideal but doesn't seem egregious.
Finally, it adds a version of the `resize` test with face-centered fields.

I also rephrased the face divB in terms of derivatives for clarity, spit out Flux-CT
device functions into a header like Face-CT has, and modified some uncommon
device-side B field functions a bit.

Finally, this adds an unrelated feature allowing re-initialization of electron temps
at user option, designed for restarting ideal simulations w/electrons enabled.