B Cleanup with GMG
The existing magnetic field cleanup invoked when resizing a
simulation uses a fork of Parthenon's old BiCGSTAB solver.
This is fine, but doesn't support SMR/AMR grids.

Parthenon's new multigrid solvers *do* support SMR grids,
So I've added a magnetic field cleanup which uses them.
As a bonus, they converge much faster and to seemingly
arbitrary precision.

The new cleanup does not support cell-centered B fields,
so the old one has been left in for now.  However,
it doesn't work in this branch due to the old solver breaking,
so it may as soon be eliminated as repaired, I'm not sure.