Better inversion floors
When floors are applied as a part of the primitive variable inversion,
this is traditionally done inconsistently: a density floor will not
reduce the internal energy, and crucially neither floor will reduce
the velocity.

Due to conservation, this isn't just a matter of which frame
the injected material uses.  Rather, it represents computing
momentum conservation given a very tiny total energy, then
giving that high velocity to a cell with the floor energy: zoom!

This is an attempt to make the floors consistent.  If the density
is floored, we recalculate the solution mu given the new density.
If the internal energy is floored, we attempt to solve with the new
energy target (though, ignoring whether we still fit rho). If a solution
with our energy does not exist, the conserved state is unphysical;
we set the velocities to zero to be minimally disruptive.

By default, unsolvable velocities are considered failures, but this
can be disabled.  If inverter floors are not for you, you may also
count any floor hit as a failure, triggering the averaging of nearby
zones.